+++ src/process_improve/experiments/__init__.py
@@ -4,6 +4,7 @@ from process_improve.experiments.analysis import analyze_experiment
 from process_improve.experiments.augment import augment_design
 from process_improve.experiments.designs import generate_design
 from process_improve.experiments.designs_factorial import full_factorial
+from process_improve.experiments.designs_omars import is_omars, omars_properties
 from process_improve.experiments.evaluate import evaluate_design
 from process_improve.experiments.factor import Constraint, DesignResult, Factor, Response, ResponseGoal
 from process_improve.experiments.knowledge import doe_knowledge
@@ -38,8 +39,10 @@ __all__ = [
     "full_factorial",
     "gather",
     "generate_design",
+    "is_omars",
     "lm",
     "main_effects_plot",
+    "omars_properties",
     "optimize_responses",
     "predict",
     "recommend_strategy",